@@ -29,10 +29,9 @@
   },
   "homepage": "https://github.com/teleport/react-dnd-treeview/",
   "devDependencies": {
-    "@types/classnames": "^0.0.31",
+    "@types/classnames": "^2.2.3",
     "@types/node": "^6.0.46",
-    "@types/react": "^0.14.44",
-    "@types/react-dnd": "^2.0.31",
+    "@types/react": "^16.4.8",
     "babel-core": "^6.0.20",
     "babel-eslint": "^4.1.3",
     "babel-loader": "^6.0.1",
@@ -49,7 +48,7 @@
   "dependencies": {
     "classnames": "^2.2.3",
     "immutable": "^3.8.1",
-    "react": "^0.14.6",
-    "react-dnd": "^2.1.3"
+    "react": "^16.4.2",
+    "react-dnd": "^3.0.2"
   }
 }
