@@ -1,16 +1,23 @@
-import React, { Component } from "react";
-import * as Immutable from "immutable";
-import { DragDropContext, Backend as DragDropBackend } from "react-dnd";
-import HTML5DragDropBackend from "react-dnd-html5-backend";
-const TouchDragDropBackend = require("react-dnd-touch-backend").default;
+import React, { Component } from 'react';
+import * as Immutable from 'immutable';
+import { DragDropContext } from 'react-dnd';
+import HTML5DragDropBackend from 'react-dnd-html5-backend';
 
-import { TreeNode, TreeNodeList, TreeNodeID, TreeView, MoveTreeNodeArgs } from "react-dnd-treeview";
+import {
+  TreeNode,
+  TreeNodeList,
+  TreeView,
+  MoveTreeNodeArgs
+} from 'react-dnd-treeview';
 
-const styles = require("./styles.css");
+declare var require: any;
+
+const styles = require('./styles.css');
 
 interface TestNode extends TreeNode {
   readonly title: string;
   readonly children?: TestNodeList;
+  parent?: TestNode;
 }
 
 interface TestNodeList extends TreeNodeList {
@@ -19,35 +26,43 @@ interface TestNodeList extends TreeNodeList {
 
 const recursivelyUpdateNode = (
   node: TestNode,
+  parentNode: TreeNode,
   listUpdateFunc: (list: TestNodeList, parentNode: TreeNode) => TestNodeList,
-  nodeUpdateFunc: (node: TestNode) => TestNode
+  nodeUpdateFunc: (node: TestNode, parentNode: TreeNode) => TestNode
 ) => {
-  const children = node.children ? node.children : { items: Immutable.List<TestNode>() };
-  const updateChildren = recursivelyUpdateList(children, node, listUpdateFunc, nodeUpdateFunc);
+  const children = node.children
+    ? node.children
+    : { items: Immutable.List<TestNode>() };
+  const updateChildren = recursivelyUpdateList(
+    children,
+    node,
+    listUpdateFunc,
+    nodeUpdateFunc
+  );
   if (updateChildren !== node.children) {
     node = Object.assign({}, node, {
-      children: updateChildren,
+      children: updateChildren
     });
   }
-  return nodeUpdateFunc(node);
+  return nodeUpdateFunc(node, parentNode);
 };
 
 const recursivelyUpdateList = (
   list: TestNodeList,
   parentNode: TreeNode,
   listUpdateFunc: (list: TestNodeList, parentNode: TreeNode) => TestNodeList,
-  nodeUpdateFunc: (node: TestNode) => TestNode
+  nodeUpdateFunc: (node: TestNode, parentNode: TreeNode) => TestNode
 ) => {
-  const mappedItems = list.items.map(item => recursivelyUpdateNode(item, listUpdateFunc, nodeUpdateFunc));
+  const mappedItems = list.items.map(item =>
+    recursivelyUpdateNode(item, parentNode, listUpdateFunc, nodeUpdateFunc)
+  );
   if (!Immutable.is(mappedItems, list.items)) {
     list = Object.assign({}, list, {
-      items: mappedItems,
+      items: mappedItems
     });
   }
   return listUpdateFunc(list, parentNode);
 };
-
-
 
 interface AppState {
   rootNodes: TestNodeList;
@@ -55,155 +70,169 @@ interface AppState {
 
 export class App extends Component<{}, AppState> {
   constructor() {
-    super();
+    super({});
     this.state = {
       rootNodes: {
         items: Immutable.List<TestNode>([
           {
-            id: "A",
-            title: "A",
+            id: 'A',
+            title: 'A',
             children: {
               items: Immutable.List<TestNode>([
                 {
-                  id: "A1",
-                  title: "A1",
+                  id: 'A1',
+                  title: 'A1'
                 },
                 {
-                  id: "A2",
-                  title: "A2",
+                  id: 'A2',
+                  title: 'A2'
                 },
                 {
-                  id: "A3",
-                  title: "A3",
-                },
-              ]),
-            },
+                  id: 'A3',
+                  title: 'A3'
+                }
+              ])
+            }
           },
           {
-            id: "B",
-            title: "B",
+            id: 'B',
+            title: 'B',
             children: {
               items: Immutable.List<TestNode>([
                 {
-                  id: "B1",
-                  title: "B1",
+                  id: 'B1',
+                  title: 'B1'
                 },
                 {
-                  id: "B2",
-                  title: "B2",
-                },
-              ]),
-            },
+                  id: 'B2',
+                  title: 'B2'
+                }
+              ])
+            }
           },
           {
-            id: "C",
-            title: "C",
+            id: 'C',
+            title: 'C',
             children: {
               items: Immutable.List<TestNode>([
                 {
-                  id: "C1",
-                  title: "C1",
+                  id: 'C1',
+                  title: 'C1',
                   children: {
                     items: Immutable.List<TestNode>([
                       {
-                        id: "C1x",
-                        title: "C1x",
+                        id: 'C1x',
+                        title: 'C1x'
                       },
                       {
-                        id: "C1y",
-                        title: "C1y",
+                        id: 'C1y',
+                        title: 'C1y'
                       },
                       {
-                        id: "C1z",
-                        title: "C1z",
+                        id: 'C1z',
+                        title: 'C1z'
                       },
                       {
-                        id: "C1zz",
-                        title: "C1zz",
+                        id: 'C1zz',
+                        title: 'C1zz'
                       },
                       {
-                        id: "C1zzz",
-                        title: "C1zzz",
-                      },
-                    ]),
-                  },
-                },
-              ]),
-            },
-          },
-        ]),
-      },
+                        id: 'C1zzz',
+                        title: 'C1zzz'
+                      }
+                    ])
+                  }
+                }
+              ])
+            }
+          }
+        ])
+      }
     };
   }
 
   handleMoveNode = (args: MoveTreeNodeArgs) => {
-    this.setState(Object.assign({}, this.state, {
-      rootNodes: recursivelyUpdateList(
-        this.state.rootNodes,
-        null,
-        (list, parentNode) =>
-          parentNode === args.newParentNode && parentNode === args.oldParentNode
-            ? Object.assign({}, list, {
-              items:
-              list.items
-                .insert(args.newParentChildIndex, args.node as TestNode)
-                .remove(args.oldParentChildIndex + (args.newParentChildIndex < args.oldParentChildIndex ? 1 : 0))
-            })
-            : parentNode === args.newParentNode
+    this.setState(
+      Object.assign({}, this.state, {
+        rootNodes: recursivelyUpdateList(
+          this.state.rootNodes,
+          null,
+          (list, parentNode) =>
+            parentNode === args.newParentNode &&
+            parentNode === args.oldParentNode
               ? Object.assign({}, list, {
-                items: list.items.insert(args.newParentChildIndex, args.node as TestNode)
-              })
-              : parentNode === args.oldParentNode
-                ? Object.assign({}, list, {
-                  items: list.items.remove(args.oldParentChildIndex)
+                  items: list.items
+                    .insert(args.newParentChildIndex, args.node as TestNode)
+                    .remove(
+                      args.oldParentChildIndex +
+                        (args.newParentChildIndex < args.oldParentChildIndex
+                          ? 1
+                          : 0)
+                    )
                 })
-                : list,
-        item => item
-      ),
-    }));
+              : parentNode === args.newParentNode
+                ? Object.assign({}, list, {
+                    items: list.items.insert(
+                      args.newParentChildIndex,
+                      args.node as TestNode
+                    )
+                  })
+                : parentNode === args.oldParentNode
+                  ? Object.assign({}, list, {
+                      items: list.items.remove(args.oldParentChildIndex)
+                    })
+                  : list,
+          item => item
+        )
+      })
+    );
   };
 
   setStateWithLog = (newState: AppState) => {
-    console.log("new state: ", newState);
+    console.log('new state: ', newState);
     this.setState(newState);
   };
 
   handleToggleCollapse = (node: TestNode) => {
-    this.setStateWithLog(Object.assign({}, this.state, {
-      rootNodes: recursivelyUpdateList(
-        this.state.rootNodes,
-        null,
-        (list, parentNode) => list,
-        item => item === node ? Object.assign({}, item, {
-          isCollapsed: !item.isCollapsed,
-        }) : item
-      ),
-    }));
+    this.setStateWithLog(
+      Object.assign({}, this.state, {
+        rootNodes: recursivelyUpdateList(
+          this.state.rootNodes,
+          null,
+          (list, parentNode) => list,
+          item =>
+            item === node
+              ? Object.assign({}, item, {
+                  isCollapsed: !item.isCollapsed
+                })
+              : item
+        )
+      })
+    );
   };
 
   renderNode = (node: TestNode) => (
-    <div className={ styles.nodeItem }>
-      { !node.children || node.children.items.isEmpty()
-        ? null
-        : <a
-          style={{ fontSize: "0.5em", verticalAlign: "middle" }}
-          onClick={ () => this.handleToggleCollapse(node) }
-          >
-          {node.isCollapsed ? "⊕" : "⊖"}
+    <div className={styles.nodeItem}>
+      {!node.children || node.children.items.isEmpty() ? null : (
+        <a
+          style={{ fontSize: '0.5em', verticalAlign: 'middle' }}
+          onClick={() => this.handleToggleCollapse(node)}
+        >
+          {node.isCollapsed ? '⊕' : '⊖'}
         </a>
-      }
-      Node: { node.title }
+      )}
+      Node: {node.title}
     </div>
   );
 
   render() {
     return (
       <TreeView
-        rootNodes={ this.state.rootNodes }
-        classNames={ styles }
-        renderNode={ this.renderNode }
-        onMoveNode={ this.handleMoveNode }
-        />
+        rootNodes={this.state.rootNodes}
+        classNames={styles}
+        renderNode={this.renderNode}
+        onMoveNode={this.handleMoveNode}
+      />
     );
   }
 }
